Register output to bus
Biwise ORs the register value with the BUS upon OutputEnable. Will not reset the BUS to 0 when OutputEnable is disabled. Cannot set bus equal to the register output as other things may be outputting to the bus simultaneously.

TODO: figure out how to reset the bus after output to the bus has been disabled without interfering with other modules outputting to the bus.